Home Page
# Description
Created a simple page with a navigation bar and some content explaining the purpose of this website.

The navigation bar includes an icon, title and a burger that on click it shows a few navigation links. They do not link to anything at the moment.

Using `gsap` to aid with animations and linked the `Roboto Condensed` font from google fonts that is being used by all text.

The svg icon used was taken from the following link: https://game-icons.net/1x1/lorc/jigsaw-box.html

The colour palette contained in the stylesheet properties was taken from the following link: https://lospec.com/palette-list/oil-6